move main components into routes folder
