+++ src/routes.ts
@@ -1,6 +1,6 @@
 import { type RouteConfig, index, route } from "@react-router/dev/routes";
 
 export default [
-  index("./components/App.tsx"),
-  route("overrides/:prefillId?", "./components/OverridesApp.tsx"),
+  index("./routes/Index.tsx"),
+  route("overrides/:prefillId?", "./routes/Overrides.tsx"),
 ] satisfies RouteConfig;
